Fix model selector lint blocker
